% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanChange.R
\name{mean_abs_change}
\alias{mean_abs_change}
\title{Mean Absolute Change}
\usage{
mean_abs_change(x)
}
\arguments{
\item{x}{a vector}
}
\description{
Mean Absolute Change
}
\examples{
\dontrun{
# don't run this sript
info <- mean_abs_change("data")
}

}
\seealso{
\code{\link{mean}}
}
