% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoWayCount.R
\name{my2waycnt}
\alias{my2waycnt}
\title{Get Two Way Count}
\usage{
my2waycnt(data, variable1, variable2)
}
\arguments{
\item{data}{a dataframe.}

\item{variable1}{a character defining the first feature.}

\item{variable2}{a character defining the second feature.}
}
\description{
Provides two way interaction count on two features, useful for feature engineering
}
\examples{
\dontrun{
# don't run this sript
2waycnt <- my2waycnt(data,"feat1","feat2")
}

}
\seealso{
\code{\link{cor} \link{freq.encode}}
}
