% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XgbModel.R
\name{XgbModel}
\alias{XgbModel}
\title{Creative Modelling with Xgboost}
\usage{
XgbModel(
  data,
  label,
  newdata,
  seed,
  folds,
  parameter,
  type = "regression",
  earlystopping = 10,
  print_every_n = 5,
  maximize = F
)
}
\arguments{
\item{data}{a dataframe or matrix denotes the training data}

\item{label}{target variable.}

\item{newdata}{a dataframe or matrix denote the test data.}

\item{seed}{numeric.}

\item{folds}{number of folds. default is 5}

\item{type}{type of model.Default = regression.}

\item{earlystopping}{numeric. default = 10}

\item{print_every_n}{numeric. default = 5}

\item{maximize}{boolean. default is F. used to maximize the metric of choice.}

\item{parameters}{list containing the model paramters.}
}
\description{
Provides a high level function to train xgboost models
}
\examples{
\dontrun{
# don't run this sript

}



}
\seealso{
\code{\link{cor} \link{detectCores}}
}
